requirements.txt
Django==4.2.13
djangorestframework==3.15.2
ipython==8.18.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dotenv==1.0.1
redis==5.0.7
celery==5.4.0
django-celery-beat==2.7.0
eventlet==0.36.1
requests==2.32.3
coverage==7.0.5
flake8==7.1.1